Style title, subtext, button and image with flexbox
I have created a flex container for all of these elements so that I would
be able to make them share space evenly. I have done this by creating
a container for the text elements and a separate container for the image.

I have aligned the text with a flex column and styled it and I have made
the image be 100% of the width of its container, with an automatic height.
I have specified a minimum width for it, so that it does not disappear
completely when the tab is minimized to an extreme amount.