Style the offer section
I have styled the offer section by including the title and the other
elements in a flexbox column, placing the title in one div and the elements
in another div.

I have centered the title by using flexbox and styled the
images by making them be divs in their own separate flexbox. I gave them
the ability to wrap and gave them a min-width so that they do not lose
their proportions when the window is made smaller.